Dispatch region formation using transform dialect (#119)
Introduce the functionality to form dispatch regions using transform
dialcet as part of openxla-transform plugin for the preprocessing stage.
This is an adaptation of https://github.com/openxla/iree/pull/11886 to
the plugin infrastructure.

It requires adding a custom interpreter pass that is capable of
constructing the transform IR on-the-fly during pass initialization in
order to correctly handle multi-threaded execution during the Flow
pipeline. A cleaner approach requires changes to upstream MLIR, which
have been landed, but take O(weeks) to propagate through IREE
integration.

The functionality is exercised using the reduction matcher already
available in IREE. However, the matcher is currently too strict to work
at the Flow level, in particular, it expects to match all tileable
operations in the parent container. Therefore the matcher code is
adapted from IREE into the plugin. This may be removed after IREE
relaxes the matcher sufficiently.

This also introduces a custom transform dialect extension with
additional operations that cannot be currently placed in IREE.